+++ qiskit/providers/ibmq/ibmqbackend.py
@@ -33,7 +33,8 @@ from qiskit.providers.ibmq.apiconstants import ApiJobShareLevel
 from .api.clients import AccountClient
 from .api.exceptions import ApiError
 from .credentials import Credentials
-from .exceptions import IBMQBackendError, IBMQBackendValueError
+from .exceptions import (IBMQBackendError, IBMQBackendValueError,
+                         IBMQBackendApiError, IBMQBackendApiProtocolError)
 from .job import IBMQJob
 from .utils import update_qobj_config
 
@@ -100,8 +101,10 @@ class IBMQBackend(BaseBackend):
 
         Raises:
             SchemaValidationError: If the job validation fails.
-            IBMQBackendError: If an unexpected error occurred while submitting
+            IBMQBackendApiError: If an unexpected error occurred while submitting
                 the job.
+            IBMQBackendApiProtocolError: If an unexpected value received when
+                 the server.
             IBMQBackendValueError: If the specified job share level is not valid.
         """
         # pylint: disable=arguments-differ
@@ -139,8 +142,12 @@ class IBMQBackend(BaseBackend):
             ibmq.job.start: The job has started.
 
         Raises:
-            IBMQBackendError: If an unexpected error occurred while submitting
+            IBMQBackendApiError: If an unexpected error occurred while submitting
                 the job.
+            IBMQBackendError: If an unexpected error occurred after submitting
+                the job.
+            IBMQBackendApiProtocolError: If an unexpected value received when
+                 the server.
         """
         try:
             qobj_dict = qobj.to_dict()
@@ -151,12 +158,13 @@ class IBMQBackend(BaseBackend):
                 job_name=job_name,
                 job_share_level=job_share_level)
         except ApiError as ex:
-            raise IBMQBackendError('Error submitting job: {}'.format(str(ex)))
+            raise IBMQBackendApiError('Error submitting job: {}'.format(str(ex)))
 
         # Error in the job after submission:
         # Transition to the `ERROR` final state.
         if 'error' in submit_info:
-            raise IBMQBackendError('Error submitting job: {}'.format(str(submit_info['error'])))
+            raise IBMQBackendError(
+                'Error submitting job: {}'.format(str(submit_info['error'])))
 
         # Submission success.
         submit_info.update({
@@ -167,8 +175,8 @@ class IBMQBackend(BaseBackend):
         try:
             job = IBMQJob.from_dict(submit_info)
         except ModelValidationError as err:
-            raise IBMQBackendError('Unexpected return value from the server when '
-                                   'submitting job: {}'.format(str(err)))
+            raise IBMQBackendApiProtocolError('Unexpected return value from the server '
+                                              'when submitting job: {}'.format(str(err)))
         Publisher().publish("ibmq.job.start", job)
         return job