+++ CHANGELOG.md
@@ -14,6 +14,16 @@ The format is based on [Keep a Changelog].
 > - **Security**: in case of vulnerabilities.
 
 
+## [UNRELEASED]
+
+### Changed
+
+- The Exception hierarchy has been refined with more specialized classes. 
+  You can, however, continue to catch their parent exceptions (such 
+  as `IBMQAccountError`). Also, the exception class `IBMQApiUrlError` 
+  has been replaced by `IBMQAccountCredentialsInvalidUrl` and 
+  `IBMQAccountCredentialsInvalidToken`. (\#480)
+
 ## [0.4.5] - 2019-12-18
 
 ### Added