+++ CHANGELOG.md
@@ -18,9 +18,25 @@ The format is based on [Keep a Changelog].
 
 ### Added
 
+- `IBMQJob` now has three new methods: `done()`, `running()`, and
+  `cancelled()`. The methods are used to indicate the job status. (\#494)
+
+### Changed
+
+- The Exception hierarchy has been refined with more specialized classes. 
+  You can, however, continue to catch their parent exceptions (such 
+  as `IBMQAccountError`). Also, the exception class `IBMQApiUrlError` 
+  has been replaced by `IBMQAccountCredentialsInvalidUrl` and 
+  `IBMQAccountCredentialsInvalidToken`. (\#480)
+
+## [0.4.5] - 2019-12-18
+
+### Added
+
 - `IBMQJob` now has a new `queue_info()` method that returns queue 
   information, such as queue position, estimated start/end time, and 
   priority levels for the job. (\#467)  
+- Python 3.8 is now supported in qiskit-ibmq-provider. (\#445)
 
 ## [0.4.4] - 2019-12-09
 
@@ -273,7 +289,8 @@ The format is based on [Keep a Changelog].
 - Support for non-qobj format has been removed. (\#26, \#28)
 
 
-[UNRELEASED]: https://github.com/Qiskit/qiskit-ibmq-provider/compare/0.4.4...HEAD
+[UNRELEASED]: https://github.com/Qiskit/qiskit-ibmq-provider/compare/0.4.5...HEAD
+[0.4.5]: https://github.com/Qiskit/qiskit-ibmq-provider/compare/0.4.4...0.4.5
 [0.4.4]: https://github.com/Qiskit/qiskit-ibmq-provider/compare/0.4.3...0.4.4
 [0.4.3]: https://github.com/Qiskit/qiskit-ibmq-provider/compare/0.4.2...0.4.3
 [0.4.2]: https://github.com/Qiskit/qiskit-ibmq-provider/compare/0.4.1...0.4.2